refactor(models,core,ui): migra monetários para Decimal e redesenha navegação
Lote 1 — Modelos: todos os campos monetários (amount, balance, limit_total,
total_amount, current_price, avg_price, etc.) migrados de float para
decimal.Decimal. Validadores Pydantic e propriedades calculadas atualizados.
Invariante: nunca Decimal(float); sempre Decimal(str) ou Decimal("literal").

Lote 2 — Core engines: analytics, behavioral, installment_engine e
weekly_report operam em Decimal end-to-end. Fronteiras com motores de
score (m1_quant, fragility, m2_governance, liquidity) usam float(inv.*)
para manter aritmética de razões/percentuais em float. sum() com acumulador
Decimal(0) em todos os geradores monetários.

UI — Navegação: st.sidebar eliminado em todas as páginas; substituído por
st.columns([1, 4]) com rail CSS. Strip operacional no Dashboard
(Patrimônio / Liquidez / Fragility / Comprometimento / Risco).
CSS nativo do Streamlit oculto; page_link customizado mantido.

Testes: 338 passando em 5 s. pytest.approx removido de asserções monetárias;
substituído por igualdade exata Decimal. test_sidebar_navigation.py adicionado.

+++ models/budget.py
@@ -1,4 +1,5 @@
 import uuid
+from decimal import Decimal
 
 from pydantic import BaseModel, Field, field_validator
 
@@ -6,13 +7,13 @@ from pydantic import BaseModel, Field, field_validator
 class Budget(BaseModel):
     id: str = Field(default_factory=lambda: str(uuid.uuid4()))
     category: str
-    monthly_limit: float
+    monthly_limit: Decimal
     year: int
     month: int
 
     @field_validator("monthly_limit")
     @classmethod
-    def limit_positive(cls, v: float) -> float:
+    def limit_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("Limite mensal deve ser positivo")
         return round(v, 2)
